Part of validation and Controller changes

+++ VM-ediaAPI/Controllers/CommentController.cs
@@ -2,6 +2,7 @@ using System;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
 using VM_ediaAPI.Data;
@@ -10,7 +11,7 @@ using VM_ediaAPI.Models;
 
 namespace VM_ediaAPI.Controllers
 {
-     [Route("api/user/{userId}/[controller]")]
+    [Route("api/user/{userId}/[controller]")]
     [ApiController]
     public class CommentController : ControllerBase
     {
@@ -22,7 +23,7 @@ namespace VM_ediaAPI.Controllers
             _repo = repo;
             _mapper = mapper;
         }
-
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult> AddComment(AddCommentDto addCommentDto)
         {
@@ -38,23 +39,41 @@ namespace VM_ediaAPI.Controllers
 
             return StatusCode(201);
         }
-
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteComment(int id)
         {
-            var comment = _repo.GetCommentById(id);
+            var comment = await _repo.GetCommentById(id);
+            if (comment == null)
+            {
+                return NotFound();
+            }
+            if (int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value) != comment.UserId)
+            {
+                return StatusCode(403);
+            }
             _repo.Delete(comment);
             await _repo.SaveAll();
             return NoContent();
         }
 
+        [Authorize]
         [HttpPatch("{id}")]
         public async Task<IActionResult> UpdateComment(int id, JsonPatchDocument<UpdateCommentDto> commentDto)
         {
             var comment = await _repo.GetCommentById(id);
+            if (comment == null)
+            {
+                return NotFound();
+            }
+            if (int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value) != comment.UserId)
+            {
+                return StatusCode(403);
+            }
             var commentToPatch = _mapper.Map<UpdateCommentDto>(comment);
             commentDto.ApplyTo(commentToPatch);
-            _repo.Edit(commentDto);
+            _mapper.Map(commentToPatch, comment);
+            _repo.Edit(comment);
             await _repo.SaveAll();
 
             return NoContent();
