+++ VM-ediaAPI/Controllers/PostController.cs
@@ -2,6 +2,7 @@ using System;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
 using VM_ediaAPI.Data;
@@ -22,7 +23,7 @@ namespace VM_ediaAPI.Controllers
             _repo = repo;
             _mapper = mapper;
         }
-
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> AddPost(AddPostDto addPostDto)
         {
@@ -53,29 +54,49 @@ namespace VM_ediaAPI.Controllers
             }
             
             var postDetails = await _repo.GetPostDetailsDto(id, userId);
+            if(postDetails == null)
+            {
+                return BadRequest("Zasób nie istnieje");
+            }
 
             return Ok(postDetails);
         }
-
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeletePost(int id)
         {
             int userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value);
             var post = await _repo.GetPostById(id);
+            if(post == null)
+            {
+                return NotFound();
+            }
             if(post.UserId == userId)
             {
                 _repo.Delete(post);
+                await _repo.SaveAll();
                 return Ok();
             }
-            return BadRequest();
+            return Unauthorized();
         }
-
+        [Authorize]
         [HttpPatch("{id}")]
         public async Task<IActionResult> UpdatePost(int id,JsonPatchDocument<UpdatePostDto> updatePostDto)
         {
-            var post = _repo.GetPostById(id);
+            var post = await _repo.GetPostById(id);
+            if(post == null)
+            {
+                return NotFound();
+            }
+            int userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value);
+            if(post.UserId !=  userId)
+            {
+                return Unauthorized();
+            }
+
             var postToPatch = _mapper.Map<UpdatePostDto>(post);
             updatePostDto.ApplyTo(postToPatch);
+            _mapper.Map(postToPatch, post);
             _repo.Edit(post);
             await _repo.SaveAll();
